Add mdns tracking service to adb host
Add a service to adb host to allow streaming of mdns discovery.
The format uses protocol buffer and include all the latest info
sent by the device (name, api, serial number).

This is a meant as replacement to "adb mdns services" which was a pull
based service that Studio called every 1 seconds when pairing.

Test: mdns_utils_test.cpp, test_device.py
Bug: 402927103
Flag: EXEMPT (adb.exe)

Change-Id: I183813c5e855caa5efd5f4527504511c49f055c5