Support incremental install for .dm files and .sdm files through adb.
.dm files are for go/art-profiles-in-play-serving and
go/art-cloud-verification. The file format was supported by adb long
ago, but it's not supported for incremental install. This CL adds the
support so that we can test incremental install for APKs with .dm files.

.sdm files are for go/art-cloud-compilation-platform-design. This is a
new file format, introduced in Android 16, to contain cloud compilation
artifacts. This CL adds the support for it so that we can test
incremental install for APKs with .sdm files.

Bug: 372868052
Test: adb install-multiple base.apk base.dm base.arm64.sdm
Flag: EXEMPT adb client only
Change-Id: I876deeebbbc0bc04f2b753c0a748e6af9320fbf8