adbd: Make adbd bind to both listen_addrs and tcp.port if set
For Android Desktop devices we use a USB "bridge" process that connects
to adbd using `localhost:5555`. We always want to have this `localhost`
bind present so that the USB bridge can always connect when adbd is
running. By always having an address present, we can avoid the
unencrypted default bind to "*:5555". This also leaves the door open to
change this to a unix domain socket in the future.

We also want to support a user calling `adb tcpip XXXX` to open up an
unencrypted port. The `adb tcpip` command sets `service.adb.tcp.port`.
The previous logic would only inspect the value of
`service.adb.tcp.port` if `listen_addrs` was empty. This change makes it
so that we respect both properties if they are set.

In theory both `listen_addrs` and `tcp.port` could be set to bind to the
same port. This would result in a binding error. We could evaluate
setting `SO_REUSEPORT` so that the different threads could bind to the
same port.

I moved the code into its own compilation unit and added a new unit test
target. We can't use a tradefed test to test this code because setting
the properties will affect the running adbd and disconnect the adb
session. The tests with the `_legacy` suffix will be deleted when the
flag is rolled out 100%.

Bug: 408058433
Test: Set both properties and verify adbd binds to all ports.
Test: Verify that the tcp.port is the one advertised on mDNS.
Test: Verify no behavior change when flag is disabled.
Test: atest adbd_unit_test
Flag: com.android.modules.adb.flags.adbd_bind_tcp_and_listen_addrs
Change-Id: Ifdbe469dcf49da508773e4f2bfa27a2d97c61f0a