Don't use new API until flag can guard it
This CL is a partial revert of aosp/3543046. We cannot use the new API
until it is properly guarded.

There is currenly no native API guard. Testing with  __builtin_available
is not enough since Framework cannot check if these events will be sent.
As a result, the only way to fix this problem in trunk-stating and bp3a
is to disable usage of the new API.

Bug: 420187719
Flag: EXEMPT (Revert)
Test: Manual
Change-Id: I146f495000b7d43d7715afe0e51db4c46518b5b1