Fix output processing of incremental install.
The read(2) system call returns as soon as any bytes are available, up
to the number of bytes requested, not necessarily the full amount asked
for.

In particular, when processing the output of incremental install, we
sometimes get "Succ" in one call and "ess" in another call. Therefore,
we should continue to call read(2) until we reach the end of the stream
or have read kMaxMessageSize bytes.

Also, this CL fixes the incremental server so that it doesn't send a
bunch of '\0's to the adb client when disconnected from abb.

Bug: 372868052
Fixes: 425381533
Bug: 425728587
Test: adb install-multiple --incremental base.apk base.dm
Flag: EXEMPT adb client only
Change-Id: Id321c15cebeb60287a2e59c72f0fb394b60d924f