Expand device discovery to CLASS_PER_INTERFACE
There is a discrepancy between the native USB backend and libusb. We
limit the class of USB device we monitor to CLASS_PER_INTERFACE which
is incorrect since some devices (such as PinePhone) use
CLASS_MISCELLANEOUS.

By using LIBUSB_HOTPLUG_MATCH_ANY in libusb hotplug, we bring it to
parity with the native backend.

Bug: 365009755
Test: Manual + already tested by CI
Flag: EXEMPT (adb.exe)
Change-Id: I360cacc2ccf191620f01d99b5cfb73ff1380c2fa