Attempt to Reconnect Only Once
When connecting to a device using MDNS, we only attempt to reconnect one time when a device disconnects. This is in contrast to we were doing previously where we attempt to reconnect every 3 seconds for one minute.

This behavior is controlled by an environment variable ADB_TCP_RECONNECT_ONCE=1

Fixes: 433981637
Bug: 74411879
Test: TODO
Change-Id: I0bda526731c7a5c89635d4bfd0c9c6e0702f049b
Flag: EXEMPT Protected by environment: ADB_TCP_RECONNECT_ONCE