Reinstate the ability to incremental install APKs without V4 signatures
This behavior was introduced by https://r.android.com/1321890, though
that CL did not do what it was described to do.

A recent ADB change (commit b813134279576a75fe3092fe969280a701e48476)
removed this behavior. If the original behavior was a feature, then the
recent change essentially broke it.

We don't know if the original original behavior is a bug or a feature,
and we'll probably never know. In any case, since this is supported by
the device, there's no real reason to reject it on the ADB side. No
matter what the original intention was, we are now adding back this
behavior and treat it as a feature.

Bug: 441479381
Test: `adb install --incremental` with an APK without a v4 signature
Flag: EXEMPT HOST_MODE

Change-Id: I34575e7730e3e3dc0a540f558d5958fd173a4ce7